convert to individual decorators - less code
